… app/classes/components/classDialog.jsx