### What?
- Improved the design of the class data presentation page for better usability and clarity.
- Fixed the alert timing by changing it from 2 seconds to 3 seconds.
- Enhanced the design of the class dialog and introduced new field input methods to improve user interaction.

### Why?
- The goal was to provide a more user-friendly experience by improving the design and presentation of the class data.
- The alert timing change ensures that users have more time to read the notifications.
- The class dialog was redesigned to allow more flexible and efficient input methods for users.

### How?
- The class page layout was adjusted, and styling improvements were made to present data more clearly.
- The alert timeout was increased from 2 seconds to 3 seconds to provide a more comfortable duration for users to read the messages.
- The class dialog was restructured, and additional input methods were added to enhance the interaction.

### Testing?
- The new design and alert timing were tested manually by interacting with the class page and observing the alert behavior.
- The updated class dialog was tested to ensure that the new input methods work correctly and the dialog is visually enhanced.

### Anything Else?
No.